Update: Formatter more test and refactors
