+++ formatter/src/main/kotlin/formatter/expression/IdentifierNodeFormatter.kt
@@ -1,6 +1,8 @@
-package formatter
+package formatter.expression
 
 import ast.AST
+import formatter.Formatter
+import formatter.rule.FormattingRule
 import kotlin.reflect.KClass
 
 class IdentifierNodeFormatter : Formatter {
