+++ formatter/src/main/kotlin/formatter/factory/FormatterMapFactory.kt
@@ -17,20 +17,20 @@ import ast.PrintLnNode
 import ast.ReadEnvNode
 import ast.ReadInputNode
 import ast.UnaryOperation
-import formatter.AssignationNodeFormatter
-import formatter.BinaryOperationNodeFormatter
-import formatter.BlockNodeFormatter
-import formatter.DeclarationNodeFormatter
 import formatter.Formatter
-import formatter.FunctionCallFormatter
-import formatter.FunctionStatementNodeFormatter
-import formatter.IdentifierNodeFormatter
-import formatter.IfStatementFormatter
-import formatter.LiteralNodeFormatter
-import formatter.PrintLnNodeFormatter
-import formatter.ReadEnvNodeFormatter
-import formatter.ReadInputNodeFormatter
-import formatter.UnaryOperationNodeFormatter
+import formatter.conditional.BlockNodeFormatter
+import formatter.conditional.IfStatementFormatter
+import formatter.expression.BinaryOperationNodeFormatter
+import formatter.expression.IdentifierNodeFormatter
+import formatter.expression.LiteralNodeFormatter
+import formatter.expression.UnaryOperationNodeFormatter
+import formatter.statement.AssignationNodeFormatter
+import formatter.statement.DeclarationNodeFormatter
+import formatter.statement.FunctionCallFormatter
+import formatter.statement.FunctionStatementNodeFormatter
+import formatter.statement.PrintLnNodeFormatter
+import formatter.statement.ReadEnvNodeFormatter
+import formatter.statement.ReadInputNodeFormatter
 import formatter.utils.DefaultOperatorFormatter
 import kotlin.reflect.KClass
