+++ formatter/src/main/kotlin/formatter/statement/AssignationNodeFormatter.kt
@@ -1,24 +1,25 @@
-package formatter
+package formatter.statement
 
 import ast.AST
-import ast.UnaryOperation
+import formatter.Formatter
+import formatter.rule.FormattingRule
 import formatter.utils.formatNextNode
 import kotlin.reflect.KClass
 
-class UnaryOperationNodeFormatter : Formatter {
+class AssignationNodeFormatter : Formatter {
     override fun format(
         node: AST,
         rule: FormattingRule,
         formatterMap: Map<KClass<out AST>, Formatter>,
     ): String {
-        val unaryOperationNode = node as UnaryOperation
+        val assignationNode = node as ast.AssignationNode
         return buildString {
-            append(unaryOperationNode.getOperator())
-            append(if (rule.hasWhiteSpaceBeforeAndAfterOperation) " " else "")
+            append(assignationNode.identifier.variableName)
+            append(if (rule.hasSpaceBetweenEqualSign) " = " else "=")
             append(
                 formatNextNode(
                     formatterMap,
-                    unaryOperationNode.getOperand(),
+                    assignationNode.expression,
                     rule,
                 ),
             )
