+++ formatter/src/main/kotlin/formatter/statement/FunctionStatementNodeFormatter.kt
@@ -1,23 +1,25 @@
-package formatter
+package formatter.statement
 
 import ast.AST
+import ast.FunctionStatementNode
+import formatter.Formatter
+import formatter.rule.FormattingRule
 import formatter.utils.formatNextNode
 import kotlin.reflect.KClass
 
-class AssignationNodeFormatter : Formatter {
+class FunctionStatementNodeFormatter : Formatter {
     override fun format(
         node: AST,
         rule: FormattingRule,
         formatterMap: Map<KClass<out AST>, Formatter>,
     ): String {
-        val assignationNode = node as ast.AssignationNode
+        val functionStatementNode = node as FunctionStatementNode
+        val function = functionStatementNode.getFunctionNode()
         return buildString {
-            append(assignationNode.identifier.variableName)
-            append(if (rule.hasSpaceBetweenEqualSign) " = " else "=")
             append(
                 formatNextNode(
                     formatterMap,
-                    assignationNode.expression,
+                    function,
                     rule,
                 ),
             )
