+++ formatter/src/main/kotlin/formatter/statement/ReadEnvNodeFormatter.kt
@@ -1,22 +1,24 @@
-package formatter
+package formatter.statement
 
 import ast.AST
+import formatter.Formatter
+import formatter.rule.FormattingRule
 import formatter.utils.formatNextNode
 import kotlin.reflect.KClass
 
-class ReadInputNodeFormatter : Formatter {
+class ReadEnvNodeFormatter : Formatter {
     override fun format(
         node: AST,
         rule: FormattingRule,
         formatterMap: Map<KClass<out AST>, Formatter>,
     ): String {
-        val readInputNode = node as ast.ReadInputNode
+        val readEnvNode = node as ast.ReadEnvNode
         return buildString {
 //            repeat(rule.lineBreakBeforePrintLn) {
 //                append("\n")
 //            }
             append(
-                "readInput(${formatNextNode(formatterMap, readInputNode.getExpression(), rule)})",
+                "readEnv(${formatNextNode(formatterMap, readEnvNode.getExpression(), rule)})",
             )
         }
     }
