+++ formatter/src/main/kotlin/formatter/statement/ReadInputNodeFormatter.kt
@@ -1,22 +1,24 @@
-package formatter
+package formatter.statement
 
 import ast.AST
+import formatter.Formatter
+import formatter.rule.FormattingRule
 import formatter.utils.formatNextNode
 import kotlin.reflect.KClass
 
-class ReadEnvNodeFormatter : Formatter {
+class ReadInputNodeFormatter : Formatter {
     override fun format(
         node: AST,
         rule: FormattingRule,
         formatterMap: Map<KClass<out AST>, Formatter>,
     ): String {
-        val readEnvNode = node as ast.ReadEnvNode
+        val readInputNode = node as ast.ReadInputNode
         return buildString {
 //            repeat(rule.lineBreakBeforePrintLn) {
 //                append("\n")
 //            }
             append(
-                "readEnv(${formatNextNode(formatterMap, readEnvNode.getExpression(), rule)})",
+                "readInput(${formatNextNode(formatterMap, readInputNode.getExpression(), rule)})",
             )
         }
     }
