+++ formatter/src/test/kotlin/conditional/IfStatementTest.kt
@@ -1,3 +1,5 @@
+package conditional
+
 import ast.BlockNode
 import ast.ColonNode
 import ast.DeclarationNode
@@ -11,68 +13,25 @@ import ast.PrintLnNode
 import ast.ProgramNode
 import ast.VariableType
 import ast.VariableTypeNode
-import formatter.FormattingRule
 import formatter.ProgramNodeFormatter
 import formatter.factory.FormatterMapFactory
+import formatter.rule.FormattingRule
 import lexer.Lexer
 import lexer.getTokenMap
-import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Assertions
 import parser.factory.ProgramParserFactory
+import parser.factory.StatementParserFactory
 import parser.result.SuccessResult
-import kotlin.test.Ignore
+import utils.getStringFromFile
+import version.Version
 import kotlin.test.Test
 
-class FormatterTest {
-    private val lexer = Lexer(getTokenMap())
+class IfStatementTest {
+    private val lexer = Lexer(getTokenMap(Version.V2))
     private val parser = ProgramParserFactory.create()
+    private val statementParser = StatementParserFactory.create(Version.V2)
     private val configPath: String = "src/test/resources/formatter.test.config.json"
 
-    @Test
-    fun `test_001 String Declaration`() {
-        val code = "let a: string=\"String\";"
-        val tokens = lexer.lex(code)
-        val result = parser.parse(tokens, 0) as SuccessResult
-        val ast = result.value as ProgramNode
-        val actual = ProgramNodeFormatter().format(ast, FormattingRule(configPath), FormatterMapFactory().createFormatterMap())
-        val expected = getStringFromFile("src/test/resources/test_001_result.ps")
-        assertEquals(expected, actual)
-    }
-
-    @Test
-    fun `test_002 Int Declaration`() {
-        val code = "let   secretNumber    : number  =           1000;"
-        val tokens = lexer.lex(code)
-        println(tokens)
-        val result = parser.parse(tokens, 0) as SuccessResult
-        val ast = result.value as ProgramNode
-        val actual = ProgramNodeFormatter().format(ast, FormattingRule(configPath), FormatterMapFactory().createFormatterMap())
-        val expected = getStringFromFile("src/test/resources/test_002_result.ps")
-        assertEquals(expected, actual)
-    }
-
-    @Test
-    fun `test_003 Two Statements`() {
-        val code = "let              first:           number =     1     ;\nlet      second    : string           = \"2\"     ;"
-        val tokens = lexer.lex(code)
-        val result = parser.parse(tokens, 0) as SuccessResult
-        val ast = result.value as ProgramNode
-        val actual = ProgramNodeFormatter().format(ast, FormattingRule(configPath), FormatterMapFactory().createFormatterMap())
-        val expected = getStringFromFile("src/test/resources/test_003_result.ps")
-        assertEquals(expected, actual)
-    }
-
-    @Test
-    fun `test_004 Print Statement`() {
-        val code = "let a          : number     =    1   ;\nprintln(1+a      )  ;"
-        val tokens = lexer.lex(code)
-        val result = parser.parse(tokens, 0) as SuccessResult
-        val ast = result.value as ProgramNode
-        val actual = ProgramNodeFormatter().format(ast, FormattingRule(configPath), FormatterMapFactory().createFormatterMap())
-        val expected = getStringFromFile("src/test/resources/test_004_result.ps")
-        assertEquals(expected, actual)
-    }
-
-//    @Ignore
     @Test
     fun `test_005 If Statement`() {
         val code = "let a: number = 1;\nif(b) {\n    println(1);\n   println(2);\n}"
@@ -139,31 +98,7 @@ class FormatterTest {
         val ast = result.value as ProgramNode
         val actual = ProgramNodeFormatter().format(ast, FormattingRule(configPath), FormatterMapFactory().createFormatterMap())
         val expected = getStringFromFile("src/test/resources/test_005_result.ps")
-        assertEquals(expected, actual)
-    }
-
-    @Ignore
-    @Test
-    fun `test_006 Const Declaration`() {
-        val code = "const a: string=\"String\";"
-        val tokens = lexer.lex(code)
-        val result = parser.parse(tokens, 0) as SuccessResult
-        val ast = result.value as ProgramNode
-        val actual = ProgramNodeFormatter().format(ast, FormattingRule(configPath), FormatterMapFactory().createFormatterMap())
-        val expected = getStringFromFile("src/test/resources/test_001_result.ps")
-        assertEquals(expected, actual)
-    }
-
-    @Ignore
-    @Test
-    fun `test_007 Boolean Declaration`() {
-        val code = "let    a:boolean =true;"
-        val tokens = lexer.lex(code)
-        val result = parser.parse(tokens, 0) as SuccessResult
-        val ast = result.value as ProgramNode
-        val actual = ProgramNodeFormatter().format(ast, FormattingRule(configPath), FormatterMapFactory().createFormatterMap())
-        val expected = getStringFromFile("src/test/resources/test_007_result.ps")
-        assertEquals(expected, actual)
+        Assertions.assertEquals(expected, actual)
     }
 
     @Test
@@ -254,7 +189,7 @@ class FormatterTest {
         val ast = result.value as ProgramNode
         val actual = ProgramNodeFormatter().format(ast, FormattingRule(configPath), FormatterMapFactory().createFormatterMap())
         val expected = getStringFromFile("src/test/resources/test_008_result.ps")
-        assertEquals(expected, actual)
+        Assertions.assertEquals(expected, actual)
     }
 
     @Test
@@ -331,7 +266,7 @@ class FormatterTest {
         val ast = result.value as ProgramNode
         val actual = ProgramNodeFormatter().format(ast, FormattingRule(configPath), FormatterMapFactory().createFormatterMap())
         val expected = getStringFromFile("src/test/resources/test_009_result.ps")
-        assertEquals(expected, actual)
+        Assertions.assertEquals(expected, actual)
     }
 
     @Test
@@ -447,6 +382,6 @@ class FormatterTest {
         val ast = result.value as ProgramNode
         val actual = ProgramNodeFormatter().format(ast, FormattingRule(configPath), FormatterMapFactory().createFormatterMap())
         val expected = getStringFromFile("src/test/resources/test_010_result.ps")
-        assertEquals(expected, actual)
+        Assertions.assertEquals(expected, actual)
     }
 }
