+++ formatter/src/test/kotlin/utils/FormatterTestUtils.kt
@@ -1,3 +1,5 @@
+package utils
+
 import java.io.File
 
 fun getStringFromFile(path: String): String {
