+++ formatter/src/test/resources/test_007_result.ps
@@ -1,1 +1,1 @@
-let a: boolean = true;
+let a : boolean = true;
