+++ formatter/src/test/resources/test_011_result.ps
@@ -1,0 +1,1 @@
+let a : number = readInput("Hola");
