+++ formatter/src/test/resources/test_012_result.ps
@@ -1,0 +1,1 @@
+let a : number = readEnv("Hola");
